requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
diff-match-patch==20230430
Django==4.2.15
django-allauth==64.0.0
django-celery==3.1.17
django-filter==24.3
django-geojson==4.1.0
django-humanize==0.1.2
django-import-export==4.1.1
django-leaflet==0.30.1
django-simple-history==3.7.0
django-tinymce==4.1.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
Faker==26.3.0
humanize==4.10.0
importlib_metadata==8.2.0
kombu==5.4.0
Markdown==3.6
openpyxl==3.1.5
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zipp==3.19.2
